print f does not provide a line break

+++ calculator.c
@@ -2,14 +2,14 @@
 #include <stdlib.h>
 int main() {
 
-    int num1;
-    int num2;
+    double num1;
+    double num2;
     printf("Enter first number: ");
-    scanf("%d", &num1);
+    scanf("%lf", &num1);
     printf("Enter second number: ");
-    scanf("%d", &num2);
+    scanf("%lf", &num2);
 
-    printf("answer: %d", num1 + num2);
+    printf("answer: %f", num1 + num2);
 
     return 0;
 }
